+++ src/main/java/com/apiculture/main/repository/StatusRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.apiculture.main.model.Status;
 
-public interface StatusRepository extends JpaRepository<Status, Short> {
+public interface StatusRepository extends JpaRepository<Status, Integer> {
 
 }
